componentize header
