+++ src/components/header-buttons/style.css
@@ -1,0 +1,24 @@
+nav a {
+  display: block;
+  height: var(--header-height);
+  line-height: var(--header-height);
+  padding: 0 15px;
+  min-width: 50px;
+  text-align: center;
+  background: var(--app-background-secondary);
+  text-decoration: none;
+  color: var(--app-color-secondary);
+  will-change: background-color;
+  cursor: pointer;
+}
+
+nav a:hover,
+nav a:active,
+nav a.active {
+  background: var(--app-background-third);
+  color: var(--app-color-third);
+}
+
+.header nav a.hidden {
+  display: none;
+}
